requirements.txt
arxiv==2.1.3
aiohttp==3.9.4
PyPDF2==3.0.1
chromadb==0.4.24
openai>=1.0.0
termcolor==2.4.0
python-dotenv==1.0.1
numpy<2.0
python-docx==0.8.11
tiktoken>=0.5.0
flask>=2.3.0
